+++ libc/generic-native/syscall/syscall.c
@@ -2,7 +2,7 @@
 #include <libc/syscall.h>
 #include <unistd.h>
 
-size_t tlu_write(int fd, const void *buf, size_t size)
+int64_t tlu_write(int fd, const void *buf, uint64_t size)
 {
 	return write(fd, buf, size);
 }
